+++ src/app/admin/dashboard/dashboard.component.ts
@@ -1,5 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { faGears } from '@fortawesome/free-solid-svg-icons';
 import { AppState } from 'models/app-state.model';
 import { GetLicense, LoadAdminSettings } from 'store/admin/admin.actions';
 import { Store } from '@ngrx/store';
@@ -7,7 +6,6 @@ import { HandleSubscriptionComponent } from 'common/handle-subscription.componen
 import { License } from 'models/settings.model';
 import { AdminService } from 'admin/admin.service';
 import { take } from 'rxjs/operators';
-import { environment } from 'environments/environment';
 
 @Component({
   selector: 'app-dashboard',
@@ -15,8 +13,6 @@ import { environment } from 'environments/environment';
   styleUrls: ['./dashboard.component.scss'],
 })
 export class DashboardComponent extends HandleSubscriptionComponent implements OnInit {
-  readonly faGears = faGears;
-  adControllerUrl = environment.adControllerUrl;
   isPanelBlocked: boolean = false;
   licenseDetailUrl: string = null;
   settings = [
@@ -25,28 +21,28 @@ export class DashboardComponent extends HandleSubscriptionComponent implements O
       description: '',
       link: '/admin/dashboard/users',
       values: [
-        { name: 'Users List', icon: 'assets/images/user-gray.svg' },
-        { name: 'Publishers', icon: 'assets/images/user-gray.svg' },
-        { name: 'Advertisers', icon: 'assets/images/user-gray.svg' },
-        { name: 'Reports', icon: 'assets/images/user-gray.svg' },
+        { name: 'Users List', path: '/all' },
+        { name: 'Publishers', path: '/publishers' },
+        { name: 'Advertisers', path: '/advertisers' },
+        { name: 'Reports', path: '/reports' },
       ],
     },
     {
       title: 'General settings',
       description: '',
       link: '/admin/dashboard/general',
-      values: [{ name: 'Bid strategy', icon: 'assets/images/preferences.svg' }],
+      values: [{ name: 'Bid strategy', path: '' }],
     },
     {
       title: 'Wallet settings',
       description: '',
       link: '/admin/dashboard/account',
       values: [
-        { name: 'Account Wallet', icon: 'assets/images/preferences.svg' },
-        { name: 'Email & password', icon: 'assets/images/preferences.svg' },
-        { name: 'Referral links', icon: 'assets/images/preferences.svg' },
-        { name: 'Access tokens', icon: 'assets/images/preferences.svg' },
-        { name: 'Newsletter', icon: 'assets/images/preferences.svg' },
+        { name: 'Account wallet', path: '/wallet' },
+        { name: 'Email & password', path: '/preferences' },
+        { name: 'Referral links', path: '/referrals' },
+        { name: 'Access tokens', path: '/access-token' },
+        { name: 'Newsletter', path: '/newsletter' },
       ],
     },
   ];