+++ src/app/advertiser/dashboard/dashboard.component.ts
@@ -15,6 +15,7 @@ import { timer } from 'rxjs';
 import { take } from 'rxjs/operators';
 import { RequestReport } from 'store/common/common.actions';
 import { reportType } from 'models/enum/user.enum';
+import { faPlusCircle } from '@fortawesome/free-solid-svg-icons';
 
 @Component({
   selector: 'app-dashboard',
@@ -36,6 +37,7 @@ export class DashboardComponent extends HandleSubscriptionComponent implements O
   barChartData = createInitialDataSet();
 
   currentChartFilterSettings: ChartFilterSettings;
+  faPlusCircle = faPlusCircle;
 
   constructor(private chartService: ChartService, private store: Store<AppState>) {
     super();