+++ src/app/advertiser/edit-campaign/edit-campaign-create-ads/edit-campaign-create-ads.component.ts
@@ -26,6 +26,8 @@ import { ADSHARES_UNITS, WIKI_HTML_AD } from 'models/enum/link.enum';
 import { Format } from 'models/taxonomy-medium.model';
 import { SessionService } from '../../../session.service';
 import { ShowDialogOnError } from 'store/common/common.actions';
+import { faPlus } from '@fortawesome/free-solid-svg-icons';
+import { faCircleXmark } from '@fortawesome/free-regular-svg-icons';
 
 interface UploadingFile {
   name: string;
@@ -75,6 +77,8 @@ export class EditCampaignCreateAdsComponent extends HandleSubscriptionComponent 
   campaign: Campaign = null;
   isEditMode: boolean;
   formats: Format[];
+  faPlus = faPlus;
+  faCircleXmark = faCircleXmark;
 
   constructor(
     private advertiserService: AdvertiserService,