+++ src/app/common/components/chart/chart-settings/chart-settings.ts
@@ -73,7 +73,7 @@ export const chartOptions = (currencyCode: string): ChartOptions<'bar'> => {
             left: position.left + 15 + tooltipModel.caretX + 'px',
             top: position.top + window.scrollY + tooltipModel.caretY + 'px',
             textAlign: 'left',
-            fontFamily: 'Lato-Regular',
+            fontFamily: 'Montserrat-Regular',
             fontSize: 14 + 'px',
             color: '#FFFFFF',
             padding: '7px 10px',
@@ -121,11 +121,11 @@ export const chartOptions = (currencyCode: string): ChartOptions<'bar'> => {
     },
     datasets: {
       bar: {
-        backgroundColor: '#55a8fd',
+        backgroundColor: '#96dfef',
         barPercentage: 0.2,
-        borderColor: '#55a8fd',
-        hoverBackgroundColor: '#4ba3fd',
-        hoverBorderColor: '#55a8fd',
+        borderColor: '#96dfef',
+        hoverBackgroundColor: '#96dfef',
+        hoverBorderColor: '#96dfef',
       },
     },
   };