+++ src/app/common/components/targeting/targeting-display/targeting-display.component.ts
@@ -2,6 +2,7 @@ import { Component, EventEmitter, Input, OnChanges, Output } from '@angular/core
 
 import { TargetingOption, TargetingOptionValue } from 'models/targeting-option.model';
 import { getPathAndLabel } from 'common/components/targeting/targeting.helpers';
+import { faClose } from '@fortawesome/free-solid-svg-icons';
 
 @Component({
   selector: 'app-targeting-display',
@@ -23,6 +24,7 @@ export class TargetingDisplayComponent implements OnChanges {
       url: string | undefined;
     }[];
   }[];
+  faClose = faClose;
 
   ngOnChanges() {
     this.prepareItemsToDisplay();