+++ src/app/impersonation/impersonation.module.ts
@@ -3,9 +3,10 @@ import { ImpersonationComponent } from './impersonation/impersonation.component'
 import { MatIconModule } from '@angular/material/icon';
 import { CommonModule } from '@angular/common';
 import { BrowserModule } from '@angular/platform-browser';
+import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 
 @NgModule({
-  imports: [CommonModule, BrowserModule, MatIconModule],
+  imports: [CommonModule, BrowserModule, MatIconModule, FontAwesomeModule],
   exports: [ImpersonationComponent],
   declarations: [ImpersonationComponent],
 })