+++ src/app/settings/reports-list-wrapper/reports-list-wrapper.component.ts
@@ -3,6 +3,5 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'app-reports-list-wrapper',
   templateUrl: './reports-list-wrapper.component.html',
-  styleUrls: ['./reports-list-wrapper.component.scss'],
 })
 export class ReportsListWrapperComponent {}