+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Displaying user id on account settings section
+- Access to AdController for moderator
+### Changed
+- Total rebuild of panel look (color palette, elements sizes, arrangement of elements)
+- Font change to Montserrat 
+- Improved navigation
+- Update labels
+- Font awesome icons instead of image assets
+### Fixed
+- External links
+- Forbidden to add parcel with incomplete coordinates
+- Click conversion choice
 
 ## [2.3.0] - 2023-01-26
 ### Added